requirements.txt
opencv-python
notebook
matplotlib
uuid
pathlib
python-dotenv